+++ src/isomedia/isom_store.c
@@ -2450,7 +2450,7 @@ GF_Err WriteToFile(GF_ISOFile *movie, Bool for_fragments)
 			gf_bs_seek(movie->editFileMap->bs, gf_bs_get_size(movie->editFileMap->bs) );
 
 			if ((movie->storageMode==GF_ISOM_STORE_FASTSTART) && mdat_start && mdat_size) {
-				u32 pad = mdat_start;
+				u32 pad = (u32) mdat_start;
 				//make sure the bitstream has the right offset - this is require for box using offsets into other boxes (typically saio)
 				moov_bs = gf_bs_new(NULL, 0, GF_BITSTREAM_WRITE);
 				while (pad) {
@@ -2489,7 +2489,7 @@ GF_Err WriteToFile(GF_ISOFile *movie, Bool for_fragments)
 				gf_bs_get_content(moov_bs, &moov_data, &moov_size);
 				gf_bs_del(moov_bs);
 				//the first mdat_start bytes are dummy, cf above
-				movie->on_block_patch(movie->on_block_out_usr_data, moov_data+mdat_start, moov_size-mdat_start, mdat_start, GF_TRUE);
+				movie->on_block_patch(movie->on_block_out_usr_data, moov_data+mdat_start, (u32) (moov_size-mdat_start), mdat_start, GF_TRUE);
 				gf_free(moov_data);
 			}
 		} else {
